requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
email-validator==1.0.4
fastapi==0.33.0
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
h11==0.8.1
httptools==0.0.13
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
line-bot-sdk==1.15.0
MarkupSafe==1.1.1
pydantic==0.30
python-dotenv==0.10.3
requests==2.22.0
starlette==0.12.0
urllib3==1.25.7
uvicorn==0.8.4
uvloop==0.12.2
websockets==7.0
Werkzeug==0.16.0
